refactor: * changed names * reorg components * deleted old `utils.ts`.
