+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import "./globals.css";
 import { Inter } from "next/font/google";
 
-import Header from "@/components/Header/Header";
-import Footer from "@/components/Footer/Footer";
+import Header from "@/components/header/Header";
+import Footer from "@/components/footer/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
