+++ components/Header/Connect.tsx
@@ -1,11 +1,11 @@
 import { WithToast } from "@/components/feedback/WithToast";
 import { Button } from "../ui/Button";
 
-export const SignIn = () => (
+export const Connect = () => (
   <WithToast
     title={"Not implemented yet"}
     subtitle={"Stay tuned as we implement walletconnect v2 soon!"}
   >
-    <Button variant={"default"}>Connect Wallet </Button>
+    <Button variant={"default"}>Connect 🔗</Button>
   </WithToast>
 );
