+++ components/Header/index.ts
@@ -1,3 +1,2 @@
 export * from "./Header";
-export * from "./SignIn";
-export * from "./Logo";
+export * from "./Connect";
