+++ components/ui/Toast.tsx
@@ -1,5 +1,5 @@
-import { useEffect } from "react";
 import { motion } from "framer-motion";
+import { useEffect } from "react";
 
 import { Button } from "@/components/ui/Button";
